DOC-6187-Config-Page-Not-Displaying-in1.5
https://issues.couchbase.com/browse/DOC-6187

In applying the Revs-Limit change to 1.5 the tabbing in the yaml file 
was inadvertently altered. This appears to have cause the page to not 
display the yaml content at all.

I have amended the yaml and ran it through the validator here:  
https://yamlvalidator.com/.

The amended page is available to view here: 
https://ibsoln.github.io/myStaging/sync-gateway/1.5/config-properties.html