Update 4.0 REST API
Running locally can cause npm to hang while executing `npx` commands.

At a guess, it's outputting a prompt "Install module?" which is being redirected, and so remains unseen.

This commit adds explicit `--yes` to run each of these without confirmation.
(It might be more elegant to set `CI=true` for the script?)